change track charge to integer to save space

+++ SkimCode/source_25/HFtrigValidation/HFtrigValidation/TrigRates.h
@@ -318,7 +318,7 @@ private:
    std::vector<float> m_muon_trk_pt ;
    std::vector<float> m_muon_trk_eta;
    std::vector<float> m_muon_trk_phi;
-   std::vector<float> m_muon_trk_charge;
+   std::vector<int>   m_muon_trk_charge;
    std::vector<int>   m_muon_trk_index;
    std::vector<float> m_muon_me_p   ;
    std::vector<int>   m_muon_elosstype;
@@ -363,7 +363,7 @@ private:
    std::vector<float> m_muon_pair_muon1_trk_pt ;
    std::vector<float> m_muon_pair_muon1_trk_eta;
    std::vector<float> m_muon_pair_muon1_trk_phi;
-   std::vector<float> m_muon_pair_muon1_trk_charge;
+   std::vector<int>   m_muon_pair_muon1_trk_charge;
    std::vector<float> m_muon_pair_muon2_pt     ;
    std::vector<float> m_muon_pair_muon2_eta    ;
    std::vector<float> m_muon_pair_muon2_phi    ;
@@ -383,7 +383,7 @@ private:
    std::vector<float> m_muon_pair_muon2_trk_pt ;
    std::vector<float> m_muon_pair_muon2_trk_eta;
    std::vector<float> m_muon_pair_muon2_trk_phi;
-   std::vector<float> m_muon_pair_muon2_trk_charge;
+   std::vector<int>   m_muon_pair_muon2_trk_charge;
